Refactoring and scene-specific compilation.
The code has been rewritten and brought into a unified style, and the names of some methods have also changed. Also, Work has been done on one part, namely, compiling the debug project to run a specific scene. Now, when you enter such debug IDs as `single` and` single_SCENE_NAME`, the project will compile the project, which will jump to the specified scene when the game starts. This is necessary if you only need to test a specific scene.